+++ packages/mesh/graphql/query/story.gql
@@ -93,6 +93,8 @@ query GetStory($storyId: ID, $picksTake: Int, $commentsTake: Int) {
     }
     content
     apiData
+    trimApiData
+    isMember
   }
 }
 
@@ -150,11 +152,9 @@ query GetStories($storyIds: [ID!], $picksTake: Int, $commentsTake: Int) {
   }
 }
 
-query GetStorySource($storyId: ID){
-  story(where:  {
-     id: $storyId
-  }){
-    source{
+query GetStorySource($storyId: ID) {
+  story(where: { id: $storyId }) {
+    source {
       id
       customId
       title
